Split out WasmRPC off of JSRPC.
