+++ test-bridge/src/main/scala/org/scalajs/testing/bridge/JSRPC.scala
@@ -14,63 +14,24 @@ package org.scalajs.testing.bridge
 
 import scala.scalajs.js
 import scala.scalajs.js.annotation._
-import scala.scalajs.LinkingInfo._
-import scala.scalajs.LinkingInfo.ModuleKind.MinimalWasmModule
-import scala.scalajs.wasm.minimal.annotation._
 
-import scala.collection.mutable
+/* Use the queue execution context (based on JS promises) explicitly:
+ * We do not have anything better at our disposal and it is accceptable in
+ * terms of fairness: JSRPC only handles in-between test communcation, so any
+ * future chain will "yield" to I/O (waiting for a message) or an RPC handler in
+ * a finite number of steps.
+ */
+import scala.scalajs.concurrent.JSExecutionContext.Implicits.queue
+
 import scala.concurrent.duration._
-import scala.concurrent.ExecutionContext
 
 import org.scalajs.testing.common.RPCCore
 
 /** JS RPC Core. Uses `scalajsCom`. */
 private[bridge] final object JSRPC extends RPCCore {
-  linkTimeIf(moduleKind == MinimalWasmModule) {
-    ()
-  } {
-    /* Use the queue execution context (based on JS promises) explicitly:
-     * We do not have anything better at our disposal and it is accceptable in
-     * terms of fairness: JSRPC only handles in-between test communication, so any
-     * future chain will "yield" to I/O (waiting for a message) or an RPC handler in
-     * a finite number of steps.
-     */
-    implicit val ec = scala.scalajs.concurrent.JSExecutionContext.queue
-    Com.init(handleMessage _)
-  }
+  Com.init(handleMessage _)
 
-  override protected def send(msg: String): Unit = {
-    linkTimeIf(moduleKind == MinimalWasmModule) {
-      WasmCom.send(stringToUTF16CodeUnits(msg))
-    } {
-      Com.send(msg)
-    }
-  }
-
-  private def stringToUTF16CodeUnits(s: String): Array[Short] = {
-    val len = s.length()
-    val codeUnits = new Array[Short](len)
-    var i = 0
-    while (i != len) {
-      codeUnits(i) = s.charAt(i).toShort
-      i += 1
-    }
-    codeUnits
-  }
-
-  @WasmExport("scalajs:testing/com/receive")
-  def receive(msg: Array[Short]): Unit = {
-    val chars = new Array[Char](msg.length)
-    var i = 0
-    while (i != chars.length) {
-      chars(i) = msg(i).toChar
-      i += 1
-    }
-
-    implicit val ec = ComLoopExecutionContext
-    handleMessage(new String(chars))
-    ec.runLoop()
-  }
+  override protected def send(msg: String): Unit = Com.send(msg)
 
   @js.native
   @JSGlobal("scalajsCom")
@@ -79,44 +40,5 @@ private[bridge] final object JSRPC extends RPCCore {
     def send(msg: String): Unit = js.native
     // We support close, but do not use it. The JS side just terminates.
     // def close(): Unit = js.native
-  }
-
-  private object WasmCom {
-    @WasmImport("scalajs:testing/com", "send")
-    def send(msg: Array[Short]): Unit = scala.scalajs.wasm.native
-
-    @WasmImport("scalajs:testing/com", "reportTopLevelError")
-    def reportTopLevelError(message: Array[Short]): Unit = scala.scalajs.wasm.native
-  }
-
-  private object ComLoopExecutionContext extends ExecutionContext {
-    private val tasks = mutable.ListBuffer.empty[Runnable]
-    private var inLoop: Boolean = false
-
-    def execute(runnable: Runnable): Unit =
-      tasks += runnable
-
-    def runLoop(): Unit = {
-      if (inLoop) {
-        // Reentrency into a loop; don't start a new one
-      } else {
-        inLoop = true
-        try {
-          while (tasks.nonEmpty) {
-            val task = tasks.remove(0)
-            try {
-              task.run()
-            } catch {
-              case t: Throwable => reportFailure(t)
-            }
-          }
-        } finally {
-          inLoop = false
-        }
-      }
-    }
-
-    def reportFailure(t: Throwable): Unit =
-      WasmCom.reportTopLevelError(stringToUTF16CodeUnits(t.toString()))
   }
 }
